벡터 이미지 JSON 형태 그대로 보존 해서 저장 벡터 이미지 검색 JSON 나오는지 확인

+++ src/main/java/com/ryu/bigdata/vo/SkuImgVo.java
@@ -7,7 +7,7 @@ public class SkuImgVo {
 
     private String vector;
     private int skuId;
-    private double skuImgId;
+    private int skuImgId;
     private String s3Path;
     private String regDateTime;
     private int numofvector;
